+++ src/main/java/fastcampus/team7/Livable_officener/dto/notification/NotificationDTO.java
@@ -1,4 +1,4 @@
-package fastcampus.team7.Livable_officener.dto;
+package fastcampus.team7.Livable_officener.dto.notification;
 
 import lombok.Data;
